Updated model loading
Have a bunch of stack allocated model containers (the mesh and index data will still be allocated on the heap, but this limits the number of models I can add).
Added a way to obtain a model's handle, and use it to render it, or to free it. I should add a way to add transformations to it.
Found a bug when adding multiple models, in which the second one uses the first's texture maps, but will look into it tomorrow.